Celery & refactoring commands
