+++ gamestore/celery.py
@@ -1,0 +1,29 @@
+import os
+from celery import Celery
+from celery.schedules import crontab
+
+# set the default Django settings module for the 'celery' program.
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'gamestore.settings')
+
+app = Celery('gamestore')
+
+# Using a string here means the worker doesn't have to serialize
+# the configuration object to child processes.
+# - namespace='CELERY' means all celery-related configuration keys
+#   should have a `CELERY_` prefix.
+app.config_from_object('django.conf:settings', namespace='CELERY')
+
+# Load task modules from all registered Django app configs.
+app.autodiscover_tasks()
+
+#
+# @app.on_after_configure.connect
+# def setup_periodic_tasks(sender, **kwargs):
+#     from board.tasks import update_favourites
+#     # Calls test('hello') every 10 minutes.
+#     sender.add_periodic_task(1*60.0, update_favourites.s(), name='add every 10m')
+#
+
+@app.task(bind=True)
+def debug_task(self):
+    print(f'Request: {self.request!r}')
